moving inventory to dedicated directory
