+++ .gitignore
@@ -51,3 +51,7 @@ coverage.xml
 
 # configs
 netbox.json
+
+# vim
+*.swo
+*.swp
